+++ Client/Scripts/Gameplay/Dto/AbilityUseProgress.cs
@@ -5,13 +5,13 @@ namespace Soteo.Gameplay.Dto;
 public sealed record AbilityUseProgress
 {
     public required AbilitySlot Slot { get; init; }
-    public float ElapsedTime { get; init; }
-    public required float RemainingTime { get; init; }
+    public double ElapsedTime { get; init; }
+    public required double RemainingTime { get; init; }
     
-    public float FullTime => ElapsedTime + RemainingTime;
-    public float NormalizedProgress => FullTime == 0 ? 1 : ElapsedTime / FullTime;
+    public double FullTime => ElapsedTime + RemainingTime;
+    public double NormalizedProgress => FullTime == 0 ? 1 : ElapsedTime / FullTime;
     
-    public AbilityUseProgress AddTime(float time)
+    public AbilityUseProgress AddTime(double time)
     {
         return this with
         {
